SASLMR

Documentation for SASLMR.

SASLMR.aov

SASLMR.aov1

SASLMR.aov2

SASLMR.aov3

SASLMR.bedata

StatsAPI.coeftable

SASLMR.aov — Method

aov(f, d; beta=true, resid=true, type = "III")

Same as aov1, aov2, aov3.

Keyword type: ANOVA SS Type I/II/III (III by default).

SASLMR.aov1 — Method

aov1(f, d; beta=true, resid=true)

Function:

aov1(Formula, Data, BETA=FALSE, Resid=FALSE)

ANOVA with Type I SS.

Arguments:

Formula a conventional formula for a linear model.

Data a data.frame to be analyzed

BETA if TRUE, coefficients (parameters) of REG will be returned. This is equivalent to SOLUTION option of SAS PROC GLM

Resid if TRUE, fitted values (y hat) and residuals will be returned

Details

It performs the core function of SAS PROC GLM, and returns Type I SS. This accepts continuous independent variables also.

Value

The result table is comparable to that of SAS PROC ANOVA.

Df degree of freedom

Sum Sq sum of square for the set of contrasts

Mean Sq mean square

F value F value for the F distribution

Pr(>F) proability of larger than F value

Next returns are optional.

Parameter Parameter table with standard error, t value, p value. TRUE is 1, and FALSE is 0 in the Estimable column. This is returned only with BETA=TRUE option.

Fitted Fitted value or y hat. This is returned only with Resid=TRUE option.

Residual Weigthed residuals. This is returned only with Resid=TRUE option.

SASLMR.aov2 — Method

aov2(f, d; beta=true, resid=true)

Function:

aov2(Formula, Data, BETA=FALSE, Resid=FALSE)

ANOVA with Type II SS.

Description see aov1.

SASLMR.aov3 — Method

aov3(f, d; beta=true, resid=true)

Function:

aov3(Formula, Data, BETA=FALSE, Resid=FALSE)

ANOVA with Type III SS.

Description see aov1.

SASLMR.bedata — Method

bedata()

Return BEdata (Contains Cmax data from a real bioequivalence study).

StatsAPI.coeftable — Method

StatsBase.coeftable(obj::AOVSumm)